now results script is clearer
